backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
bcrypt
pydantic
python-multipart
pandas
openpyxl
python-jose[cryptography]
passlib[bcrypt]